Move authentication to a separate module
