+++ app/routes.py
@@ -23,7 +23,7 @@ from app import (
     app,
     db,
 )
-from app.email import send_password_reset_email
+from app.auth.email import send_password_reset_email
 from app.forms import (
     EditProfileForm,
     EmptyForm,
@@ -109,57 +109,6 @@ def explore():
         next_url=url_for("explore", page=posts.next_num) if posts.has_next else None,
         prev_url=url_for("explore", page=posts.prev_num) if posts.has_prev else None,
     )
-
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for("index"))
-
-    form = LoginForm()
-
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user is None or not user.check_password(form.password.data):
-            flash(_("Invalid username or password"))
-            return redirect(url_for("login"))
-
-        login_user(user, remember=form.remember_me.data)
-
-        next = request.args.get("next")
-        if not next or url_parse(next).netloc != "":
-            next = url_for("index")
-
-        return redirect(next)
-
-    return render_template("login.html", title=_("Sign In"), form=form)
-
-
-@app.route("/logout")
-def logout():
-    logout_user()
-    return redirect(url_for("index"))
-
-
-@app.route("/register", methods=["GET", "POST"])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for("index"))
-
-    form = RegistrationForm()
-
-    if form.validate_on_submit():
-        user = User(username=form.username.data, email=form.email.data)
-        user.set_password(form.password.data)
-
-        db.session.add(user)
-        db.session.commit()
-
-        flash(_("Congratulations, you are now a registered user!"))
-
-        return redirect(url_for("login"))
-
-    return render_template("register.html", title=_("Register"), form=form)
 
 
 @app.route("/user/<username>")
@@ -272,49 +221,6 @@ def unfollow(username):
         return redirect(url_for("user", username=username))
     else:
         return redirect(url_for("index"))
-
-
-@app.route("/reset_password_request", methods=["GET", "POST"])
-def reset_password_request():
-    if current_user.is_authenticated:
-        return redirect(url_for("index"))
-
-    form = ResetPasswordRequestForm()
-
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-
-        if user:
-            send_password_reset_email()
-
-        flash(_("Check your email for the instructions to reset your password"))
-        return redirect(url_for("login"))
-
-    return render_template(
-        "reset_password_request.html", title=_("Reset Password"), form=form
-    )
-
-
-@app.route("/reset_password/<token>", methods=["GET", "POST"])
-def reset_password(token: str):
-    if current_user.is_authenticated:
-        return redirect(url_for("index"))
-
-    user = User.verify_reset_password_token(token)
-    if not user:
-        redirect(url_for("index"))
-
-    form = ResetPasswordForm()
-
-    if form.validate_on_submit():
-        user.set_password(form.password.data)
-
-        db.session.commit()
-
-        flash(_("Your password has been reset."))
-        return redirect(url_for("login"))
-
-    return render_template("reset_password.html", form=form)
 
 
 @app.route("/translate", methods=["POST"])
