+++ app/templates/auth/reset_password.html
@@ -2,7 +2,7 @@
 {% import "bootstrap/wtf.html" as wtf %}
 
 {% block app_content %}
-    <h1>{{ _("Reset Your Password") }}</h1>
+    <h1>{{ _("Reset Password") }}</h1>
 
     <div class="row">
         <div class="col-md-4">
